Moved mirage node fixtures to conftest for reusability
- Relocated mirage_active_nodes and mirage_passive_nodes fixtures from test file to conftest
- Added required import for generate_random_node_data utility
